chore(ci): update dev dependencies
This updates commitlint, semantic-release, and minimist to address a security issue with older versions of minimist. Commitlint and semantic-release should be fine because their breaking changes were updates to Node 10 which we already use. Minimist doesn't have a changelog so I have no idea what might break. :) All of these dependencies are only used for doing automated releases in CI so they shouldn't affect the actual functionality of this package.